PR fixes: naming, javadocs
