+++ channels/jaicp/src/main/kotlin/com/justai/jaicf/channel/jaicp/execution/ThreadPoolRequestExecutor.kt
@@ -94,7 +94,7 @@ private fun tryProcessAsExternalInvocation(
     } catch (e: Exception) {
         return false
     }
-    channel.processExternalInvocation(
+    channel.processInvocation(
         request = InvocationEventRequest(data.chatId, event, request.raw),
         requestContext = RequestContext.fromHttp(request.asHttpBotRequest())
     )
