+++ channels/jaicp/src/main/kotlin/com/justai/jaicf/channel/jaicp/reactions/JaicpAsyncReactions.kt
@@ -5,9 +5,8 @@ import com.justai.jaicf.channel.jaicp.http.ChatAdapterConnector
 import com.justai.jaicf.channel.jaicp.livechat.LiveChatInitRequest
 import com.justai.jaicf.channel.jaicp.livechat.exceptions.NoOperatorChannelConfiguredException
 import com.justai.jaicf.channel.jaicp.livechat.exceptions.NoOperatorsOnlineException
-import com.justai.jaicf.logging.Reaction
+import com.justai.jaicf.channel.jaicp.reactions.reaction.SwitchReaction
 import com.justai.jaicf.reactions.jaicp.JaicpCompatibleAsyncReactions
-import kotlinx.serialization.json.JsonObject
 
 /**
  * Switches to livechat operator if channel is connected to livechat in JAICP App Console.
@@ -37,36 +36,4 @@ fun JaicpCompatibleAsyncReactions.switchToLiveChat(reply: LiveChatSwitchReply): 
     connector.initLiveChat(switchRequest)
     return SwitchReaction.fromReply(switchRequest.switchData, loggingContext.botContext.dialogContext.currentState)
         .also { loggingContext.reactions.add(it) }
-}
-
-data class SwitchReaction(
-    val firstMessage: String? = null,
-    val closeChatPhrases: List<String> = emptyList(),
-    val appendCloseChatButton: Boolean = false,
-    val ignoreOffline: Boolean = false,
-    val oneTimeMessage: Boolean = false,
-    val destination: String? = null,
-    val lastMessage: String? = null,
-    val attributes: JsonObject? = null,
-    val hiddenAttributes: JsonObject? = null,
-    val sendMessagesToOperator: Boolean = false,
-    val sendMessageHistoryAmount: Int? = null,
-    override val fromState: String
-) : Reaction(fromState) {
-    companion object {
-        fun fromReply(switchReply: LiveChatSwitchReply, state: String) = SwitchReaction(
-            firstMessage = switchReply.firstMessage,
-            closeChatPhrases = switchReply.closeChatPhrases,
-            appendCloseChatButton = switchReply.appendCloseChatButton,
-            ignoreOffline = switchReply.ignoreOffline,
-            oneTimeMessage = switchReply.oneTimeMessage,
-            destination = switchReply.destination,
-            lastMessage = switchReply.lastMessage,
-            attributes = switchReply.attributes,
-            hiddenAttributes = switchReply.hiddenAttributes,
-            sendMessageHistoryAmount = switchReply.sendMessageHistoryAmount,
-            sendMessagesToOperator = switchReply.sendMessagesToOperator,
-            fromState = state
-        )
-    }
 }
